Refactor code to make it more feasible to dynamically add indexes without changing so much code. This changes to a generic header for the table of contents, and the resulting section would contain a long which must uniquely identify the section type.
There is still a missing way to dynamically register index types, but
there is now an entry point in the form of a generic map that
maps from the index type to the code.

currently this is limited to the ColumnFileWriter, so changes would
need to make the extension points visible when constructing
ColumnFileWriter objects in callers.